iot_edge_ids/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
paho-mqtt==1.6.1
influxdb-client==1.40.0
